Move dockerutil package under internal directory
The dockerutil package is not intended to be part of a public API.

And add package comment.
