+++ chain/penumbra/penumbra_chain.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"github.com/strangelove-ventures/ibctest/chain/tendermint"
-	"github.com/strangelove-ventures/ibctest/dockerutil"
 	"github.com/strangelove-ventures/ibctest/ibc"
+	"github.com/strangelove-ventures/ibctest/internal/dockerutil"
 	"github.com/strangelove-ventures/ibctest/test"
 
 	"github.com/ory/dockertest/v3"
