+++ internal/dockerutil/doc.go
@@ -1,0 +1,2 @@
+// Package dockerutil contains helpers for interacting with Docker containers.
+package dockerutil
